Move all rescript legacy js sources to rescript-legacy folder
