+++ cli/rescript-legacy/bsb.js
@@ -6,7 +6,7 @@ import { createServer } from "node:http";
 import * as os from "node:os";
 import * as path from "node:path";
 
-import { rescript_legacy_exe } from "./bins.js";
+import { rescript_legacy_exe } from "../common/bins.js";
 import { WebSocket } from "./minisocket.js";
 
 const cwd = process.cwd();
