+++ cli/rescript-legacy/format.js
@@ -8,7 +8,7 @@ import * as os from "node:os";
 import * as path from "node:path";
 import { promisify } from "node:util";
 
-import * as arg from "#cli/args";
+import * as arg from "./args.js";
 
 const asyncExecFile = promisify(child_process.execFile);
