+++ packages/artifacts.json
@@ -8,13 +8,13 @@
     "README.md",
     "cli/bsc.js",
     "cli/bstracing.js",
-    "cli/common/args.js",
     "cli/common/bins.js",
-    "cli/common/bsb.js",
-    "cli/common/minisocket.js",
     "cli/rescript-legacy.js",
+    "cli/rescript-legacy/args.js",
+    "cli/rescript-legacy/bsb.js",
     "cli/rescript-legacy/dump.js",
     "cli/rescript-legacy/format.js",
+    "cli/rescript-legacy/minisocket.js",
     "cli/rescript-tools.js",
     "cli/rescript.js",
     "docs/docson/build-schema.json",
